working vec i64

+++ contracts/hello_world/src/lib.rs
@@ -42,13 +42,13 @@ pub struct SimpleStorageContract;
 
 #[contractimpl]
 impl SimpleStorageContract {
-    pub fn store_value(env: Env, key: Symbol, value: Vec<(i32, i32)>) {
+    pub fn store_value(env: Env, key: Symbol, value: Vec<(i64, i64)>) {
         // let map = vec![&env, (2, 20)];
         // env.storage().persistent().set(&key, &map);
         env.storage().persistent().set(&key, &value);
     }
 
-    pub fn retrieve_value(env: Env, key: Symbol) -> Option<Vec<(i32, i32)>> {
+    pub fn retrieve_value(env: Env, key: Symbol) -> Option<Vec<(i64, i64)>> {
         env.storage().persistent().get(&key)
     }
 }
